requirements.txt
python-dotenv
streamlit
streamlit_chat
langchain 


